+++ src/ColorScheme/COLOR_SCHEME.mjs
@@ -1,3 +1,5 @@
+export const COLOR_SCHEME_DARK = "dark";
+
 export const COLOR_SCHEME_LIGHT = "light";
 
-export const COLOR_SCHEME_DARK = "dark";
+export const COLOR_SCHEME_SYSTEM = "system";