+++ src/ColorScheme/ColorScheme.mjs
@@ -1,6 +1,3 @@
-/** @typedef {import("./COLOR_SCHEME.mjs").COLOR_SCHEME_DARK} COLOR_SCHEME_DARK */
-/** @typedef {import("./COLOR_SCHEME.mjs").COLOR_SCHEME_LIGHT} COLOR_SCHEME_LIGHT */
-
 /**
- * @typedef {{color_scheme: COLOR_SCHEME_LIGHT | COLOR_SCHEME_DARK, name: string}} ColorScheme
+ * @typedef {{"color-scheme"?: string | null, label: string, name: string}} ColorScheme
  */