+++ src/ColorScheme/ColorSchemeWithSystemColorScheme.mjs
@@ -1,5 +1,6 @@
 /** @typedef {import("./ColorScheme.mjs").ColorScheme} ColorScheme */
+/** @typedef {import("./SystemColorScheme.mjs").SystemColorScheme} SystemColorScheme */
 
 /**
- * @typedef {ColorScheme & {system_color_scheme: boolean}} ColorSchemeWithSystemColorScheme
+ * @typedef {ColorScheme & {system_color_scheme: SystemColorScheme | null}} ColorSchemeWithSystemColorScheme
  */