+++ src/ColorScheme/FluxSelectColorSchemeElement.mjs
@@ -1,0 +1,223 @@
+import { COLOR_SCHEME_CSS_PROPERTY_PREFIX } from "./COLOR_SCHEME_CSS_PROPERTY_PREFIX.mjs";
+import { COLOR_SCHEME_LOCALIZATION_MODULE } from "../Localization/_LOCALIZATION_MODULE.mjs";
+import { COLOR_SCHEME_SYSTEM } from "./COLOR_SCHEME.mjs";
+import { flux_css_api } from "../../../flux-css-api/src/FluxCssApi.mjs";
+import { COLOR_SCHEME_VARIABLE_ACCENT, COLOR_SCHEME_VARIABLE_BACKGROUND, COLOR_SCHEME_VARIABLE_FOREGROUND } from "./COLOR_SCHEME_VARIABLE.mjs";
+
+/** @typedef {import("./ColorScheme.mjs").ColorScheme} ColorScheme */
+/** @typedef {import("../../../flux-localization-api/src/FluxLocalizationApi.mjs").FluxLocalizationApi} FluxLocalizationApi */
+/** @typedef {import("./SystemColorScheme.mjs").SystemColorScheme} SystemColorScheme */
+
+const variables_css = await flux_css_api.import(
+    `${import.meta.url.substring(0, import.meta.url.lastIndexOf("/"))}/FluxSelectColorSchemeElementVariables.css`
+);
+
+document.adoptedStyleSheets.unshift(variables_css);
+
+const css = await flux_css_api.import(
+    `${import.meta.url.substring(0, import.meta.url.lastIndexOf("/"))}/FluxSelectColorSchemeElement.css`
+);
+
+export class FluxSelectColorSchemeElement extends HTMLElement {
+    /**
+     * @type {FluxLocalizationApi}
+     */
+    #flux_localization_api;
+    /**
+     * @type {ShadowRoot}
+     */
+    #shadow;
+    /**
+     * @type {boolean}
+     */
+    #show_color_scheme_accent_color;
+    /**
+     * @type {SystemColorScheme[] | null}
+     */
+    #system_color_schemes;
+
+    /**
+     * @param {ColorScheme[]} color_schemes
+     * @param {SystemColorScheme[] | null} system_color_schemes
+     * @param {[ColorScheme, {[key: string]: ColorScheme}]} settings
+     * @param {(settings: [ColorScheme, {[key: string]: ColorScheme}]) => Promise<void>} setSettings
+     * @param {boolean} show_color_scheme_accent_color
+     * @param {FluxLocalizationApi} flux_localization_api
+     * @returns {Promise<FluxSelectColorSchemeElement>}
+     */
+    static async new(color_schemes, system_color_schemes, settings, setSettings, show_color_scheme_accent_color, flux_localization_api) {
+        const flux_select_color_scheme_element = new this(
+            system_color_schemes,
+            show_color_scheme_accent_color,
+            flux_localization_api
+        );
+
+        const color_schemes_element = document.createElement("div");
+        color_schemes_element.classList.add("color_schemes");
+        if (flux_select_color_scheme_element.#show_color_scheme_accent_color) {
+            color_schemes_element.dataset.show_accent_color = true;
+        }
+        await flux_select_color_scheme_element.#render(
+            color_schemes_element,
+            color_schemes,
+            settings,
+            async color_scheme => {
+                settings[0] = color_scheme;
+
+                await setSettings(
+                    settings
+                );
+            },
+            setSettings
+        );
+        flux_select_color_scheme_element.#shadow.appendChild(color_schemes_element);
+
+        return flux_select_color_scheme_element;
+    }
+
+    /**
+     * @param {SystemColorScheme[] | null} system_color_schemes
+     * @param {boolean} show_color_scheme_accent_color
+     * @param {FluxLocalizationApi} flux_localization_api
+     * @private
+     */
+    constructor(system_color_schemes, show_color_scheme_accent_color, flux_localization_api) {
+        super();
+
+        this.#system_color_schemes = system_color_schemes;
+        this.#show_color_scheme_accent_color = show_color_scheme_accent_color;
+        this.#flux_localization_api = flux_localization_api;
+
+        this.#shadow = this.attachShadow({
+            mode: "closed"
+        });
+
+        this.#shadow.adoptedStyleSheets.push(css);
+    }
+
+    /**
+     * @param {HTMLDivElement} parent_element
+     * @param {ColorScheme[]} color_schemes
+     * @param {[ColorScheme, {[key: string]: ColorScheme}]} settings
+     * @param {(color_scheme: ColorScheme) => Promise<void>} setColorScheme
+     * @param {(settings: [ColorScheme, {[key: string]: ColorScheme}]) => Promise<void>} setSettings
+     * @returns {Promise<void>}
+     */
+    async #render(parent_element, color_schemes, settings, setColorScheme, setSettings) {
+        for (const color_scheme of color_schemes) {
+            const color_scheme_element = document.createElement("button");
+            color_scheme_element.classList.add("color_scheme");
+            if (color_scheme.name === settings[0].name) {
+                color_scheme_element.dataset.selected = true;
+            }
+            color_scheme_element.title = await this.#flux_localization_api.translate(
+                color_scheme.label,
+                COLOR_SCHEME_LOCALIZATION_MODULE
+            );
+            color_scheme_element.type = "button";
+
+            if (color_scheme.name === COLOR_SCHEME_SYSTEM) {
+                color_scheme_element.dataset.system = true;
+            } else {
+                for (const variable of [
+                    ...this.#show_color_scheme_accent_color ? [
+                        COLOR_SCHEME_VARIABLE_ACCENT
+                    ] : [],
+                    COLOR_SCHEME_VARIABLE_BACKGROUND,
+                    COLOR_SCHEME_VARIABLE_FOREGROUND
+                ]) {
+                    color_scheme_element.style.setProperty(`--flux-select-color-scheme-color-scheme-${variable}-color`, `var(${COLOR_SCHEME_CSS_PROPERTY_PREFIX}${color_scheme.name}-${variable})`);
+                }
+            }
+
+            color_scheme_element.addEventListener("click", async () => {
+                if (color_scheme_element.dataset.selected === "true") {
+                    return;
+                }
+
+                parent_element.querySelectorAll(".color_scheme[data-selected]").forEach(_color_scheme_element => {
+                    delete _color_scheme_element.dataset.selected;
+                });
+
+                color_scheme_element.dataset.selected = true;
+
+                await setColorScheme(
+                    color_scheme
+                );
+
+                await this.#updateSystemColorSchemeSelector(
+                    parent_element,
+                    color_schemes,
+                    settings,
+                    setSettings
+                );
+            });
+
+            parent_element.appendChild(color_scheme_element);
+        }
+
+        await this.#updateSystemColorSchemeSelector(
+            parent_element,
+            color_schemes,
+            settings,
+            setSettings
+        );
+    }
+
+    /**
+     * @param {HTMLDivElement} parent_element
+     * @param {ColorScheme[]} color_schemes
+     * @param {[ColorScheme, {[key: string]: ColorScheme}]} settings
+     * @param {(settings: [ColorScheme, {[key: string]: ColorScheme}]) => Promise<void>} setSettings
+     * @returns {Promise<void>}
+     */
+    async #updateSystemColorSchemeSelector(parent_element, color_schemes, settings, setSettings) {
+        if (this.#system_color_schemes === null) {
+            return;
+        }
+
+        Array.from(parent_element.querySelectorAll("[data-system_selector]")).forEach(system_selector_element => {
+            system_selector_element.remove();
+        });
+
+        if (settings[0].name !== COLOR_SCHEME_SYSTEM) {
+            return;
+        }
+
+        for (const system_color_scheme of this.#system_color_schemes) {
+            const title_element = document.createElement("div");
+            title_element.classList.add("title");
+            title_element.dataset.system_selector = true;
+            title_element.innerText = await this.#flux_localization_api.translate(
+                system_color_scheme.label,
+                COLOR_SCHEME_LOCALIZATION_MODULE
+            );
+            parent_element.appendChild(title_element);
+
+            const color_schemes_element = document.createElement("div");
+            color_schemes_element.classList.add("color_schemes");
+            color_schemes_element.dataset.system_selector = true;
+            await this.#render(
+                color_schemes_element,
+                color_schemes.filter(color_scheme => color_scheme.name !== COLOR_SCHEME_SYSTEM),
+                [
+                    settings[1][system_color_scheme.name],
+                    {}
+                ],
+                async color_scheme => {
+                    settings[1][system_color_scheme.name] = color_scheme;
+
+                    await setSettings(
+                        settings
+                    );
+                },
+                setSettings
+            );
+            parent_element.appendChild(color_schemes_element);
+        }
+    }
+}
+
+export const SELECT_COLOR_SCHEME_ELEMENT_TAG_NAME = "flux-select-color-scheme";
+
+customElements.define(SELECT_COLOR_SCHEME_ELEMENT_TAG_NAME, FluxSelectColorSchemeElement);