+++ src/ColorScheme/SystemColorScheme.mjs
@@ -1,6 +1,3 @@
-/** @typedef {import("./COLOR_SCHEME.mjs").COLOR_SCHEME_DARK} COLOR_SCHEME_DARK */
-/** @typedef {import("./COLOR_SCHEME.mjs").COLOR_SCHEME_LIGHT} COLOR_SCHEME_LIGHT */
-
 /**
- * @typedef {{[key: string]: string, [COLOR_SCHEME_LIGHT]: string, [COLOR_SCHEME_DARK]: string}} SystemColorScheme
+ * @typedef {{"default-color-scheme": string, detector: MediaQueryList, label: string, name: string}} SystemColorScheme
  */