+++ src/ColorScheme/setColorScheme.mjs
@@ -1,3 +1,0 @@
-/**
- * @typedef {(color_scheme: string) => void} setColorScheme
- */