+++ src/FluxColorScheme.mjs
@@ -1,32 +1,37 @@
 import { COLOR_SCHEME_CSS_PROPERTY_PREFIX } from "./ColorScheme/COLOR_SCHEME_CSS_PROPERTY_PREFIX.mjs";
 import { COLOR_SCHEME_LOCALIZATION_MODULE } from "./Localization/_LOCALIZATION_MODULE.mjs";
-import { COLOR_SCHEME_SETTINGS_KEY } from "./Settings/COLOR_SCHEME_SETTINGS_KEY.mjs";
 import { flux_css_api } from "../../flux-css-api/src/FluxCssApi.mjs";
-import { COLOR_SCHEME_DARK, COLOR_SCHEME_LIGHT } from "./ColorScheme/COLOR_SCHEME.mjs";
-import { VARIABLE_ACCENT, VARIABLE_ACCENT_FOREGROUND, VARIABLE_ACCENT_FOREGROUND_RGB, VARIABLE_ACCENT_RGB, VARIABLE_BACKGROUND, VARIABLE_BACKGROUND_RGB, VARIABLE_FOREGROUND, VARIABLE_FOREGROUND_RGB } from "./ColorScheme/VARIABLE.mjs";
+import { COLOR_SCHEME_LIGHT, COLOR_SCHEME_SYSTEM } from "./ColorScheme/COLOR_SCHEME.mjs";
+import { COLOR_SCHEME_SETTINGS_KEY, COLOR_SCHEME_SYSTEM_SETTINGS_KEY } from "./Settings/COLOR_SCHEME_SETTINGS_KEY.mjs";
+import { COLOR_SCHEME_VARIABLE_ACCENT, COLOR_SCHEME_VARIABLE_ACCENT_FOREGROUND, COLOR_SCHEME_VARIABLE_ACCENT_FOREGROUND_RGB, COLOR_SCHEME_VARIABLE_ACCENT_RGB, COLOR_SCHEME_VARIABLE_BACKGROUND, COLOR_SCHEME_VARIABLE_BACKGROUND_RGB, COLOR_SCHEME_VARIABLE_FOREGROUND, COLOR_SCHEME_VARIABLE_FOREGROUND_RGB } from "./ColorScheme/COLOR_SCHEME_VARIABLE.mjs";
+import { DEFAULT_COLOR_SCHEMES, DEFAULT_SYSTEM_COLOR_SCHEMES } from "./ColorScheme/DEFAULT_COLOR_SCHEMES.mjs";
 
 /** @typedef {import("./ColorScheme/ColorScheme.mjs").ColorScheme} ColorScheme */
 /** @typedef {import("./ColorScheme/ColorSchemeWithSystemColorScheme.mjs").ColorSchemeWithSystemColorScheme} ColorSchemeWithSystemColorScheme */
 /** @typedef {import("../../flux-localization-api/src/FluxLocalizationApi.mjs").FluxLocalizationApi} FluxLocalizationApi */
+/** @typedef {import("./ColorScheme/FluxSelectColorSchemeElement.mjs").FluxSelectColorSchemeElement} FluxSelectColorSchemeElement */
 /** @typedef {import("../../flux-settings-api/src/FluxSettingsApi.mjs").FluxSettingsApi} FluxSettingsApi */
-/** @typedef {import("./ColorScheme/SelectColorSchemeElement.mjs").SelectColorSchemeElement} SelectColorSchemeElement */
 /** @typedef {import("./ColorScheme/SystemColorScheme.mjs").SystemColorScheme} SystemColorScheme */
 
 const variables_css = await flux_css_api.import(
-    `${import.meta.url.substring(0, import.meta.url.lastIndexOf("/"))}/ColorScheme/ColorSchemeVariables.css`
+    `${import.meta.url.substring(0, import.meta.url.lastIndexOf("/"))}/ColorScheme/FluxColorSchemeVariables.css`
 );
 
 document.adoptedStyleSheets.unshift(variables_css);
 
 export class FluxColorScheme {
     /**
-     * @type {string[] | null}
+     * @type {string[]}
      */
     #additional_variables;
     /**
      * @type {ColorScheme[]}
      */
     #color_schemes;
+    /**
+     * @type {string}
+     */
+    #default_color_scheme;
     /**
      * @type {FluxLocalizationApi | null}
      */
@@ -36,11 +41,19 @@ export class FluxColorScheme {
      */
     #flux_settings_api;
     /**
-     * @type {MediaQueryList | null}
+     * @type {() => {} | null}
      */
-    #system_color_scheme_detector = null;
+    #init_system_color_scheme_detectors = null;
     /**
-     * @type {SystemColorScheme | null}
+     * @type {boolean}
+     */
+    #set_system_color_schemes;
+    /**
+     * @type {boolean}
+     */
+    #show_color_scheme_accent_color;
+    /**
+     * @type {SystemColorScheme[]}
      */
     #system_color_schemes;
     /**
@@ -49,37 +62,55 @@ export class FluxColorScheme {
     #variables_style_sheet_rule = null;
 
     /**
-     * @param {ColorScheme[]} color_schemes
+     * @param {ColorScheme[] | null} color_schemes
+     * @param {string | null} default_color_scheme
+     * @param {string[] | null} additional_variables
+     * @param {SystemColorScheme[] | null} system_color_schemes
+     * @param {boolean | null} set_system_color_schemes
+     * @param {boolean | null} show_color_scheme_accent_color
      * @param {FluxLocalizationApi | null} flux_localization_api
      * @param {FluxSettingsApi | null} flux_settings_api
-     * @param {SystemColorScheme | null} system_color_schemes
-     * @param {string[] | null} additional_variables
-     * @returns {FluxColorScheme}
+     * @returns {Promise<FluxColorScheme>}
      */
-    static new(color_schemes, flux_localization_api = null, flux_settings_api = null, system_color_schemes = null, additional_variables = null) {
-        return new this(
-            color_schemes,
+    static async new(color_schemes, default_color_scheme = null, additional_variables = null, system_color_schemes = null, set_system_color_schemes = null, show_color_scheme_accent_color = null, flux_localization_api = null, flux_settings_api = null) {
+        const _color_schemes = color_schemes ?? DEFAULT_COLOR_SCHEMES;
+
+        const flux_color_scheme = new this(
+            _color_schemes,
+            default_color_scheme ?? (_color_schemes.some(color_scheme => color_scheme.name === COLOR_SCHEME_SYSTEM) ? null : _color_schemes[0]?.name ?? null) ?? COLOR_SCHEME_SYSTEM,
+            additional_variables ?? [],
+            system_color_schemes ?? DEFAULT_SYSTEM_COLOR_SCHEMES,
+            set_system_color_schemes ?? false,
+            show_color_scheme_accent_color ?? false,
             flux_localization_api,
-            flux_settings_api,
-            system_color_schemes,
-            additional_variables
+            flux_settings_api
         );
+
+        await flux_color_scheme.#render();
+
+        return flux_color_scheme;
     }
 
     /**
      * @param {ColorScheme[]} color_schemes
+     * @param {string} default_color_scheme
+     * @param {string[]} additional_variables
+     * @param {SystemColorScheme[]} system_color_schemes
+     * @param {boolean} set_system_color_schemes
+     * @param {boolean} show_color_scheme_accent_color
      * @param {FluxLocalizationApi | null} flux_localization_api
      * @param {FluxSettingsApi | null} flux_settings_api
-     * @param {SystemColorScheme | null} system_color_schemes
-     * @param {string[] | null} additional_variables
      * @private
      */
-    constructor(color_schemes, flux_localization_api, flux_settings_api, system_color_schemes, additional_variables) {
+    constructor(color_schemes, default_color_scheme, additional_variables, system_color_schemes, set_system_color_schemes, show_color_scheme_accent_color, flux_localization_api, flux_settings_api) {
         this.#color_schemes = color_schemes;
+        this.#default_color_scheme = default_color_scheme;
+        this.#additional_variables = additional_variables;
+        this.#system_color_schemes = system_color_schemes;
+        this.#set_system_color_schemes = set_system_color_schemes;
+        this.#show_color_scheme_accent_color = show_color_scheme_accent_color;
         this.#flux_localization_api = flux_localization_api;
         this.#flux_settings_api = flux_settings_api;
-        this.#system_color_schemes = system_color_schemes;
-        this.#additional_variables = additional_variables;
 
         if (this.#flux_localization_api !== null) {
             this.#flux_localization_api.addModule(
@@ -87,8 +118,6 @@ export class FluxColorScheme {
                 COLOR_SCHEME_LOCALIZATION_MODULE
             );
         }
-
-        this.renderColorScheme();
     }
 
     /**
@@ -96,7 +125,7 @@ export class FluxColorScheme {
      */
     async getAccent() {
         return this.getVariable(
-            VARIABLE_ACCENT
+            COLOR_SCHEME_VARIABLE_ACCENT
         );
     }
 
@@ -105,7 +134,7 @@ export class FluxColorScheme {
      */
     async getAccentForeground() {
         return this.getVariable(
-            VARIABLE_ACCENT_FOREGROUND
+            COLOR_SCHEME_VARIABLE_ACCENT_FOREGROUND
         );
     }
 
@@ -114,7 +143,7 @@ export class FluxColorScheme {
      */
     async getAccentForegroundRgb() {
         return this.getVariable(
-            VARIABLE_ACCENT_FOREGROUND_RGB
+            COLOR_SCHEME_VARIABLE_ACCENT_FOREGROUND_RGB
         );
     }
 
@@ -123,7 +152,7 @@ export class FluxColorScheme {
      */
     async getAccentRgb() {
         return this.getVariable(
-            VARIABLE_ACCENT_RGB
+            COLOR_SCHEME_VARIABLE_ACCENT_RGB
         );
     }
 
@@ -132,7 +161,7 @@ export class FluxColorScheme {
      */
     async getBackground() {
         return this.getVariable(
-            VARIABLE_BACKGROUND
+            COLOR_SCHEME_VARIABLE_BACKGROUND
         );
     }
 
@@ -141,7 +170,7 @@ export class FluxColorScheme {
      */
     async getBackgroundRgb() {
         return this.getVariable(
-            VARIABLE_BACKGROUND_RGB
+            COLOR_SCHEME_VARIABLE_BACKGROUND_RGB
         );
     }
 
@@ -149,21 +178,36 @@ export class FluxColorScheme {
      * @returns {Promise<ColorSchemeWithSystemColorScheme>}
      */
     async getColorScheme() {
-        let color_scheme_name = await this.#flux_settings_api?.get(
-            COLOR_SCHEME_SETTINGS_KEY
-        ) ?? "";
+        const [
+            color_scheme,
+            system
+        ] = await this.#getSettings();
 
-        const system_color_scheme = this.#system_color_schemes !== null && color_scheme_name === "";
-        if (system_color_scheme) {
-            color_scheme_name = ((await this.#getSystemColorSchemeDetector()).matches ? this.#system_color_schemes[COLOR_SCHEME_DARK] : this.#system_color_schemes[COLOR_SCHEME_LIGHT]) ?? "";
+        const is_system_color_scheme = color_scheme.name === COLOR_SCHEME_SYSTEM;
+
+        await this.#initSystemColorSchemesDetectors(
+            is_system_color_scheme
+        );
+
+        let _color_scheme = color_scheme;
+        let system_color_scheme = null;
+
+        if (is_system_color_scheme) {
+            _color_scheme = null;
+
+            system_color_scheme = await this.#getCurrentSystemColorScheme();
+
+            if (system_color_scheme !== null && (system[system_color_scheme.name] ?? null) !== null) {
+                _color_scheme = system[system_color_scheme.name];
+            }
+
+            if (_color_scheme === null) {
+                throw new Error("Invalid color scheme");
+            }
         }
 
-        const color_scheme = this.#color_schemes.find(_color_scheme => _color_scheme.name === color_scheme_name) ?? this.#color_schemes[0] ?? {};
-
         return {
-            color_scheme: "",
-            name: "",
-            ...color_scheme,
+            ..._color_scheme,
             system_color_scheme
         };
     }
@@ -173,7 +217,7 @@ export class FluxColorScheme {
      */
     async getForeground() {
         return this.getVariable(
-            VARIABLE_FOREGROUND
+            COLOR_SCHEME_VARIABLE_FOREGROUND
         );
     }
 
@@ -182,31 +226,29 @@ export class FluxColorScheme {
      */
     async getForegroundRgb() {
         return this.getVariable(
-            VARIABLE_FOREGROUND_RGB
+            COLOR_SCHEME_VARIABLE_FOREGROUND_RGB
         );
     }
 
     /**
-     * @returns {Promise<SelectColorSchemeElement>}
+     * @returns {Promise<FluxSelectColorSchemeElement>}
      */
     async getSelectColorSchemeElement() {
         if (this.#flux_localization_api === null) {
             throw new Error("Missing FluxLocalizationApi");
         }
 
-        return (await import("./ColorScheme/SelectColorSchemeElement.mjs")).SelectColorSchemeElement.new(
-            await this.getColorScheme(),
+        return (await import("./ColorScheme/FluxSelectColorSchemeElement.mjs")).FluxSelectColorSchemeElement.new(
             this.#color_schemes,
-            this.#flux_localization_api,
-            color_scheme_name => {
-                this.#setColorScheme(
-                    color_scheme_name
+            this.#set_system_color_schemes ? this.#system_color_schemes : null,
+            await this.#getSettings(),
+            async settings => {
+                await this.#setSettings(
+                    settings
                 );
             },
-            await this.#getVariables(
-                true
-            ),
-            this.#system_color_schemes
+            this.#show_color_scheme_accent_color,
+            this.#flux_localization_api
         );
     }
 
@@ -219,71 +261,99 @@ export class FluxColorScheme {
     }
 
     /**
-     * @param {boolean | null} only_if_system_color_scheme
-     * @returns {Promise<void>}
+     * @param {string} name
+     * @returns {Promise<ColorScheme | null>}
      */
-    async renderColorScheme(only_if_system_color_scheme = null) {
-        const color_scheme = await this.getColorScheme();
-
-        if ((only_if_system_color_scheme ?? false) && !color_scheme.system_color_scheme) {
-            return;
-        }
-
-        const variables_style_sheet_rule = await this.#getVariablesStyleSheetRule();
-        for (const key of Array.from(variables_style_sheet_rule.style).filter(_key => _key.startsWith(COLOR_SCHEME_CSS_PROPERTY_PREFIX))) {
-            variables_style_sheet_rule.style.removeProperty(key);
-        }
-        for (const variable of await this.#getVariables()) {
-            variables_style_sheet_rule.style.setProperty(`${COLOR_SCHEME_CSS_PROPERTY_PREFIX}${variable}`, `var(${COLOR_SCHEME_CSS_PROPERTY_PREFIX}${color_scheme.name}-${variable})`);
-        }
-
-        const color_scheme_meta = document.head.querySelector("meta[name=color-scheme]") ?? document.createElement("meta");
-        color_scheme_meta.content = color_scheme.color_scheme !== "" ? color_scheme.color_scheme : COLOR_SCHEME_LIGHT;
-        color_scheme_meta.name = "color-scheme";
-        if (!color_scheme_meta.isConnected) {
-            document.head.appendChild(color_scheme_meta);
-        }
-
-        const theme_color_meta = document.head.querySelector("meta[name=theme-color]") ?? document.createElement("meta");
-        theme_color_meta.content = await this.getAccent();
-        theme_color_meta.name = "theme-color";
-        if (!theme_color_meta.isConnected) {
-            document.head.appendChild(theme_color_meta);
-        }
+    async #getColorSchemeByName(name) {
+        return this.#color_schemes.find(color_scheme => color_scheme.name === name) ?? null;
     }
 
     /**
-     * @returns {Promise<MediaQueryList>}
+     * @returns {Promise<SystemColorScheme | null>}
      */
-    async #getSystemColorSchemeDetector() {
-        if (this.#system_color_scheme_detector === null) {
-            this.#system_color_scheme_detector ??= matchMedia(`(prefers-color-scheme:${COLOR_SCHEME_DARK})`);
-
-            this.#system_color_scheme_detector.addEventListener("change", () => {
-                this.renderColorScheme(
-                    true
-                );
-            });
-        }
-
-        return this.#system_color_scheme_detector;
+    async #getCurrentSystemColorScheme() {
+        return this.#system_color_schemes.find(system_color_scheme => system_color_scheme.detector.matches) ?? null;
     }
 
     /**
-     * @param {boolean | null} only_default
+     * @returns {Promise<[ColorScheme, {[key: string]: ColorScheme}]>}
+     */
+    async #getSettings() {
+        let color_scheme = null;
+
+        const name = await this.#flux_settings_api?.get(
+            COLOR_SCHEME_SETTINGS_KEY
+        ) ?? null;
+
+        if (name !== null) {
+            color_scheme = await this.#getColorSchemeByName(
+                name
+            );
+        }
+
+        if (color_scheme === null) {
+            color_scheme = await this.#getColorSchemeByName(
+                this.#default_color_scheme
+            );
+        }
+
+        if (color_scheme === null) {
+            throw new Error("Invalid color scheme");
+        }
+
+        const system = this.#set_system_color_schemes ? await this.#flux_settings_api?.get(
+            COLOR_SCHEME_SYSTEM_SETTINGS_KEY
+        ) ?? null : null;
+
+        return [
+            color_scheme,
+            Object.fromEntries(await Promise.all(this.#system_color_schemes.map(async system_color_scheme => {
+                let _color_scheme = null;
+
+                if ((system?.[system_color_scheme.name] ?? null) !== null) {
+                    _color_scheme = await this.#getColorSchemeByName(
+                        system[system_color_scheme.name]
+                    );
+                }
+
+                if (_color_scheme === null) {
+                    _color_scheme = await this.#getColorSchemeByName(
+                        system_color_scheme["default-color-scheme"]
+                    );
+                }
+
+                if (_color_scheme === null && this.#default_color_scheme !== COLOR_SCHEME_SYSTEM) {
+                    _color_scheme = await this.#getColorSchemeByName(
+                        this.#default_color_scheme
+                    );
+                }
+
+                if (_color_scheme === null) {
+                    throw new Error("Invalid color scheme");
+                }
+
+                return [
+                    system_color_scheme.name,
+                    _color_scheme
+                ];
+            })))
+        ];
+    }
+
+    /**
      * @returns {Promise<string[]>}
      */
-    async #getVariables(only_default = null) {
+    async #getVariables() {
         return Array.from(new Set([
-            VARIABLE_ACCENT,
-            VARIABLE_ACCENT_FOREGROUND,
-            VARIABLE_ACCENT_FOREGROUND_RGB,
-            VARIABLE_ACCENT_RGB,
-            VARIABLE_BACKGROUND,
-            VARIABLE_BACKGROUND_RGB,
-            VARIABLE_FOREGROUND,
-            VARIABLE_FOREGROUND_RGB,
-            ...(!(only_default ?? false) ? this.#additional_variables : null) ?? []
+            COLOR_SCHEME_VARIABLE_ACCENT,
+            COLOR_SCHEME_VARIABLE_ACCENT_FOREGROUND,
+            COLOR_SCHEME_VARIABLE_ACCENT_FOREGROUND_RGB,
+            COLOR_SCHEME_VARIABLE_ACCENT_RGB,
+            COLOR_SCHEME_VARIABLE_BACKGROUND,
+            COLOR_SCHEME_VARIABLE_BACKGROUND_RGB,
+            COLOR_SCHEME_VARIABLE_FOREGROUND,
+            COLOR_SCHEME_VARIABLE_FOREGROUND_RGB,
+            ...this.#additional_variables
         ]));
     }
 
@@ -312,19 +382,95 @@ export class FluxColorScheme {
     }
 
     /**
-     * @param {string} color_scheme_name
+     * @param {boolean} is_system_color_scheme
      * @returns {Promise<void>}
      */
-    async #setColorScheme(color_scheme_name) {
+    async #initSystemColorSchemesDetectors(is_system_color_scheme) {
+        if (is_system_color_scheme) {
+            if (this.#init_system_color_scheme_detectors !== null) {
+                return;
+            }
+
+            this.#init_system_color_scheme_detectors = () => {
+                this.#render();
+            };
+
+            for (const system_color_scheme of this.#system_color_schemes) {
+                system_color_scheme.detector.addEventListener("change", this.#init_system_color_scheme_detectors);
+            }
+        } else {
+            if (this.#init_system_color_scheme_detectors === null) {
+                return;
+            }
+
+            for (const system_color_scheme of this.#system_color_schemes) {
+                system_color_scheme.detector.removeEventListener("change", this.#init_system_color_scheme_detectors);
+            }
+
+            this.#init_system_color_scheme_detectors = null;
+        }
+    }
+
+    /**
+     * @returns {Promise<void>}
+     */
+    async #render() {
+        const color_scheme = await this.getColorScheme();
+
+        const variables_style_sheet_rule = await this.#getVariablesStyleSheetRule();
+        for (const key of Array.from(variables_style_sheet_rule.style).filter(_key => _key.startsWith(COLOR_SCHEME_CSS_PROPERTY_PREFIX))) {
+            variables_style_sheet_rule.style.removeProperty(key);
+        }
+        for (const variable of await this.#getVariables()) {
+            variables_style_sheet_rule.style.setProperty(`${COLOR_SCHEME_CSS_PROPERTY_PREFIX}${variable}`, `var(${COLOR_SCHEME_CSS_PROPERTY_PREFIX}${color_scheme.name}-${variable})`);
+        }
+
+        const color_scheme_meta_element = document.head.querySelector("meta[name=color-scheme]") ?? document.createElement("meta");
+        color_scheme_meta_element.content = color_scheme["color-scheme"] ?? COLOR_SCHEME_LIGHT;
+        color_scheme_meta_element.name = "color-scheme";
+        if (!color_scheme_meta_element.isConnected) {
+            document.head.appendChild(color_scheme_meta_element);
+        }
+
+        const theme_color_meta_element = document.head.querySelector("meta[name=theme-color]") ?? document.createElement("meta");
+        theme_color_meta_element.content = await this.getAccent();
+        theme_color_meta_element.name = "theme-color";
+        if (!theme_color_meta_element.isConnected) {
+            document.head.appendChild(theme_color_meta_element);
+        }
+    }
+
+    /**
+     * @param {[ColorScheme, {[key: string]: ColorScheme}]} settings
+     * @returns {Promise<void>}
+     */
+    async #setSettings(settings) {
         if (this.#flux_settings_api === null) {
             throw new Error("Missing FluxSettingsApi");
         }
 
         await this.#flux_settings_api.store(
             COLOR_SCHEME_SETTINGS_KEY,
-            color_scheme_name
+            settings[0].name
         );
 
-        await this.renderColorScheme();
+        if (this.#set_system_color_schemes) {
+            await this.#flux_settings_api.store(
+                COLOR_SCHEME_SYSTEM_SETTINGS_KEY,
+                Object.fromEntries(Object.entries(settings[1]).map(([
+                    name,
+                    color_scheme
+                ]) => [
+                        name,
+                        color_scheme.name
+                    ]))
+            );
+        } else {
+            await this.#flux_settings_api.delete(
+                COLOR_SCHEME_SYSTEM_SETTINGS_KEY
+            );
+        }
+
+        await this.#render();
     }
 }