+++ src/Settings/COLOR_SCHEME_SETTINGS_KEY.mjs
@@ -1,1 +1,3 @@
 export const COLOR_SCHEME_SETTINGS_KEY = "color-scheme";
+
+export const COLOR_SCHEME_SYSTEM_SETTINGS_KEY = "color-scheme-system";